Add feed validation to Follow requests from Web App and API to avoid invalid feeds
