+++ internal/handlers.go
@@ -1,7 +1,6 @@
 package internal
 
 import (
-	"bufio"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -33,7 +32,6 @@ import (
 	"github.com/vcraescu/go-paginator/adapter"
 	"gopkg.in/yaml.v2"
 
-	"github.com/prologic/twtxt"
 	"github.com/prologic/twtxt/internal/session"
 	"github.com/prologic/twtxt/types"
 )
@@ -1344,258 +1342,6 @@ func (s *Server) LookupHandler() httprouter.Handle {
 
 		w.Header().Set("Content-Type", "application/json")
 		w.Write(data)
-	}
-}
-
-// FollowHandler ...
-func (s *Server) FollowHandler() httprouter.Handle {
-	return func(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
-		ctx := NewContext(s.config, s.db, r)
-
-		nick := strings.TrimSpace(r.FormValue("nick"))
-		url := NormalizeURL(r.FormValue("url"))
-
-		if r.Method == "GET" && nick == "" && url == "" {
-			ctx.Title = "Follow a new feed"
-			s.render("follow", w, ctx)
-			return
-		}
-
-		if nick == "" || url == "" {
-			ctx.Error = true
-			ctx.Message = "Both nick and url must be specified"
-			s.render("error", w, ctx)
-			return
-		}
-
-		user := ctx.User
-		if user == nil {
-			log.Fatalf("user not found in context")
-			return
-		}
-
-		user.Following[nick] = url
-
-		if err := s.db.SetUser(ctx.Username, user); err != nil {
-			ctx.Error = true
-			ctx.Message = fmt.Sprintf("Error following feed %s: %s", nick, url)
-			s.render("error", w, ctx)
-			return
-		}
-
-		if strings.HasPrefix(url, s.config.BaseURL) {
-			url = UserURL(url)
-			nick := NormalizeUsername(filepath.Base(url))
-
-			if s.db.HasUser(nick) {
-				followee, err := s.db.GetUser(nick)
-				if err != nil {
-					log.WithError(err).Errorf("error loading user object for %s", nick)
-					ctx.Error = true
-					ctx.Message = "Error following user"
-					s.render("error", w, ctx)
-					return
-				}
-
-				if followee.Followers == nil {
-					followee.Followers = make(map[string]string)
-				}
-
-				followee.Followers[user.Username] = user.URL
-
-				if err := s.db.SetUser(followee.Username, followee); err != nil {
-					log.WithError(err).Warnf("error updating user object for followee %s", followee.Username)
-					ctx.Error = true
-					ctx.Message = "Error following user"
-					s.render("error", w, ctx)
-					return
-				}
-
-				if _, err := AppendSpecial(
-					s.config, s.db,
-					twtxtBot,
-					fmt.Sprintf(
-						"FOLLOW: @<%s %s> from @<%s %s> using %s/%s",
-						followee.Username, URLForUser(s.config, followee.Username),
-						user.Username, URLForUser(s.config, user.Username),
-						"twtxt", twtxt.FullVersion(),
-					),
-				); err != nil {
-					log.WithError(err).Warnf("error appending special FOLLOW post")
-				}
-			} else if s.db.HasFeed(nick) {
-				feed, err := s.db.GetFeed(nick)
-				if err != nil {
-					log.WithError(err).Errorf("error loading feed object for %s", nick)
-					ctx.Error = true
-					ctx.Message = "Error following user"
-					s.render("error", w, ctx)
-					return
-				}
-
-				feed.Followers[user.Username] = user.URL
-
-				if err := s.db.SetFeed(feed.Name, feed); err != nil {
-					log.WithError(err).Warnf("error updating user object for followee %s", feed.Name)
-					ctx.Error = true
-					ctx.Message = "Error following feed"
-					s.render("error", w, ctx)
-					return
-				}
-
-				if _, err := AppendSpecial(
-					s.config, s.db,
-					twtxtBot,
-					fmt.Sprintf(
-						"FOLLOW: @<%s %s> from @<%s %s> using %s/%s",
-						feed.Name, URLForUser(s.config, feed.Name),
-						user.Username, URLForUser(s.config, user.Username),
-						"twtxt", twtxt.FullVersion(),
-					),
-				); err != nil {
-					log.WithError(err).Warnf("error appending special FOLLOW post")
-				}
-			}
-		}
-
-		ctx.Error = false
-		ctx.Message = fmt.Sprintf("Successfully started following %s: %s", nick, url)
-		s.render("error", w, ctx)
-		return
-	}
-}
-
-// ImportHandler ...
-func (s *Server) ImportHandler() httprouter.Handle {
-	return func(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
-		ctx := NewContext(s.config, s.db, r)
-
-		if r.Method == "GET" {
-			ctx.Title = "Import feeds from a list"
-			s.render("import", w, ctx)
-			return
-		}
-
-		feeds := r.FormValue("feeds")
-
-		if feeds == "" {
-			ctx.Error = true
-			ctx.Message = "Nothing to import!"
-			s.render("error", w, ctx)
-			return
-		}
-
-		user := ctx.User
-		if user == nil {
-			log.Fatalf("user not found in context")
-		}
-
-		re := regexp.MustCompile(`(?P<nick>.*?)[: ](?P<url>.*)`)
-
-		imported := 0
-
-		scanner := bufio.NewScanner(strings.NewReader(feeds))
-		for scanner.Scan() {
-			line := scanner.Text()
-			matches := re.FindStringSubmatch(line)
-			if len(matches) == 3 {
-				nick := strings.TrimSpace(matches[1])
-				url := NormalizeURL(strings.TrimSpace(matches[2]))
-				if nick != "" && url != "" {
-					user.Following[nick] = url
-					imported++
-				}
-			}
-		}
-		if err := scanner.Err(); err != nil {
-			log.WithError(err).Error("error scanning feeds for import")
-			ctx.Error = true
-			ctx.Message = "Error importing feeds"
-			s.render("error", w, ctx)
-		}
-
-		if err := s.db.SetUser(ctx.Username, user); err != nil {
-			ctx.Error = true
-			ctx.Message = "Error importing feeds"
-			s.render("error", w, ctx)
-			return
-		}
-
-		ctx.Error = false
-		ctx.Message = fmt.Sprintf("Successfully imported %d feeds", imported)
-		s.render("error", w, ctx)
-		return
-	}
-}
-
-// UnfollowHandler ...
-func (s *Server) UnfollowHandler() httprouter.Handle {
-	return func(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
-		ctx := NewContext(s.config, s.db, r)
-
-		nick := strings.TrimSpace(r.FormValue("nick"))
-
-		if nick == "" {
-			ctx.Error = true
-			ctx.Message = "No nick specified to unfollow"
-			s.render("error", w, ctx)
-			return
-		}
-
-		user := ctx.User
-		if user == nil {
-			log.Fatalf("user not found in context")
-		}
-
-		url, ok := user.Following[nick]
-		if !ok {
-			ctx.Error = true
-			ctx.Message = fmt.Sprintf("No feed found by the nick %s", nick)
-			s.render("error", w, ctx)
-			return
-		}
-
-		delete(user.Following, nick)
-
-		if err := s.db.SetUser(ctx.Username, user); err != nil {
-			ctx.Error = true
-			ctx.Message = fmt.Sprintf("Error unfollowing feed %s: %s", nick, url)
-			s.render("error", w, ctx)
-			return
-		}
-
-		if strings.HasPrefix(url, s.config.BaseURL) {
-			url = UserURL(url)
-			nick := NormalizeUsername(filepath.Base(url))
-			followee, err := s.db.GetUser(nick)
-			if err != nil {
-				log.WithError(err).Warnf("error loading user object for followee %s", nick)
-			} else {
-				if followee.Followers != nil {
-					delete(followee.Followers, user.Username)
-					if err := s.db.SetUser(followee.Username, followee); err != nil {
-						log.WithError(err).Warnf("error updating user object for followee %s", followee.Username)
-					}
-				}
-				if _, err := AppendSpecial(
-					s.config, s.db,
-					twtxtBot,
-					fmt.Sprintf(
-						"UNFOLLOW: @<%s %s> from @<%s %s> using %s/%s",
-						followee.Username, URLForUser(s.config, followee.Username),
-						user.Username, URLForUser(s.config, user.Username),
-						"twtxt", twtxt.FullVersion(),
-					),
-				); err != nil {
-					log.WithError(err).Warnf("error appending special FOLLOW post")
-				}
-			}
-		}
-
-		ctx.Error = false
-		ctx.Message = fmt.Sprintf("Successfully stopped following %s: %s", nick, url)
-		s.render("error", w, ctx)
-		return
 	}
 }
